chore(Tests): moved rules tests in their own folder
