+++ tests/rules/forbid-lowercase-jsx-tags.spec.ts
@@ -1,5 +1,5 @@
 import { ESLintUtils } from "@typescript-eslint/utils";
-import rule from "../lib/rules/forbid-lowercase-jsx-tags";
+import rule from "../../lib/rules/forbid-lowercase-jsx-tags";
 
 const ruleTester = new ESLintUtils.RuleTester({
   parser: "@typescript-eslint/parser",
