+++ tests/rules/mutation-decorator-type-mismatch.spec.ts
@@ -1,5 +1,5 @@
 import { ESLintUtils } from "@typescript-eslint/utils";
-import rule from "../lib/rules/mutation-decorator-return-type-mismatch";
+import rule from "../../lib/rules/mutation-decorator-return-type-mismatch";
 
 const ruleTester = new ESLintUtils.RuleTester({
   parser: "@typescript-eslint/parser",
