+++ tests/rules/no-key-or-ref-prop.spec.ts
@@ -1,5 +1,5 @@
 import { ESLintUtils } from "@typescript-eslint/utils";
-import rule from "../lib/rules/no-key-or-ref-prop";
+import rule from "../../lib/rules/no-key-or-ref-prop";
 
 const ruleTester = new ESLintUtils.RuleTester({
   parser: "@typescript-eslint/parser",
