+++ tests/rules/no-value-export-in-declaration-file.spec.ts
@@ -1,5 +1,5 @@
 import { ESLintUtils } from "@typescript-eslint/utils";
-import rule from "../lib/rules/no-value-export-in-declaration-file";
+import rule from "../../lib/rules/no-value-export-in-declaration-file";
 
 const ruleTesterDeclarationFile = new ESLintUtils.RuleTester({
   parser: "@typescript-eslint/parser",
