+++ tests/rules/property-decorator-type-mismatch.spec.ts
@@ -1,5 +1,5 @@
 import { ESLintUtils } from "@typescript-eslint/utils";
-import rule from "../lib/rules/property-decorator-type-mismatch";
+import rule from "../../lib/rules/property-decorator-type-mismatch";
 
 const ruleTester = new ESLintUtils.RuleTester({
   parser: "@typescript-eslint/parser",
