+++ tests/rules/redux-saga-no-sequential-actions.spec.ts
@@ -1,5 +1,5 @@
 import { ESLintUtils } from "@typescript-eslint/utils";
-import rule from "../lib/rules/redux-saga-no-sequential-actions";
+import rule from "../../lib/rules/redux-saga-no-sequential-actions";
 
 const ruleTester = new ESLintUtils.RuleTester({
     parser: "@typescript-eslint/parser",
