+++ tests/rules/typeorm-enforce-repository-generic-method-typing.spec.ts
@@ -1,5 +1,5 @@
 import { ESLintUtils } from "@typescript-eslint/utils";
-import rule from "../lib/rules/typeorm-enforce-repository-generic-method-typing";
+import rule from "../../lib/rules/typeorm-enforce-repository-generic-method-typing";
 
 const ruleTester = new ESLintUtils.RuleTester({
   parser: "@typescript-eslint/parser",
