+++ tests/rules/typeorm-query-runner-release.spec.ts
@@ -1,5 +1,5 @@
 import { ESLintUtils } from "@typescript-eslint/utils";
-import rule from "../lib/rules/typeorm-query-runner-release";
+import rule from "../../lib/rules/typeorm-query-runner-release";
 
 const ruleTester = new ESLintUtils.RuleTester({
   parser: "@typescript-eslint/parser",
